+++ framework/widgets/ActiveField.php
@@ -417,6 +417,7 @@ class ActiveField extends Component
         $options = array_merge($this->inputOptions, $options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeHiddenInput($this->model, $this->attribute, $options);
+        $this->parts['{label}'] = '';
 
         return $this;
     }