+++ framework/CHANGELOG.md
@@ -48,6 +48,7 @@ Yii Framework 2 Change Log
 - Bug #13307: Preventing of race conditions in script filter in `yii.js` works incorrectly (arogachev)
 - Bug #13310: Handle relative and absolute URLs coincidence in CSS filter in `yii.js` (arogachev)
 - Bug #13312: `skipOuterContainers` option is incorrectly passed to pjax in `handleAction` in `yii.js` (arogachev)
+- Bug #12969: Improved unique ID generation for `yii\widgets\Pjax` widgets (dynasource, samdark, rob006)
 - Enh #475: Added Bash and Zsh completion support for the `./yii` command (cebe, silverfire)
 - Enh #6242: Access to validator in inline validation (arogachev)
 - Enh #6373: Introduce `yii\db\Query::emulateExecution()` to force returning an empty result for a query (klimov-paul)
@@ -105,6 +106,7 @@ Yii Framework 2 Change Log
 -----------------------
 
 - Bug #7670: Added `yii\web\UrlNormalizer` for normalizing requests with and without trailing slashes (rob006, cronfy, klimov-paul)
+- Bug #12345 Fixed `Formatter::asCurrency()` for proper decimal formatting (Oxyaction)
 - Bug #7670: Added `UrlNormalizer` for normalizing requests with and without trailing slashes (rob006, cronfy, klimov-paul)
 - Bug #9027: Fixed descendant class of `yii\web\UploadedFile` returns parent instances in case invoked after it (andrewnester)
 - Bug #9277: Fixed `yii\console\controllers\AssetController` looses custom options of 'target' bundles (petrabarus, klimov-paul)